+++ RELEASE_NOTES.md
@@ -42,6 +42,9 @@ a term position while verifying that the position is for the given term.
 dynamic binding cache update operations with mutexes when using multi-threaded
 backends. Thanks to Lindsey Spratt for help in debugging the race conditions.
 
+* FIXED: Workaround mishandling of multifile and dynamic predicates by some
+backends such as ECLiPSe and Trealla Prolog.
+
 Documentation
 -------------